Add length-based fecundity estimate data file
Draft version of length-based (i.e., time-varying) estimated fecundity by population, year, and age, for returning adult females. For missing values (population/year/age groups with zero length measurements, identifiable in this data set as records where [n_lengths] is `NA`), the average length for that population and age across all years is used instead of a year-specific mean.

The best approach for imputation of missing values, as well as whether imputation should also be applied to means that are currently based on very small sample size, is  still to be discussed internally by ODFW staff, and the data file may be updated following that discussion.