Update 06-juv-survival-hydro.csv
Replace juvenile hydrosystem survival data set (06-juv-survival) with an updated version of the data set that continues through migration year 2021 (brood year 2019). Note that as of April 2023, the Comparative Survival Study has not yet published survival estimates for the 2022 outmigration cohort (brood year 2020), either in their 2022 annual report or in the online query tool.  Note also that this revised data set also adds values for migration year 2001, which are available through the online query but are (for unknown reasons) excluded from the tables (Table A1) published in CSS reports. Finally, note that this updated data set also revises many of the previous values for MY 2018 and 2019, which were explicitly noted in the published 2020 report table as subject to change "as groups are finalized for estimation of SARs".