Update adult-abundance.csv
Correct [n_returned] for some early Minam years, where the calculations for expanding redd counts had been applied incorrectly.
The value for 1993 was not corrected because a different redd count and expansion method was used in this year. Also the redd count and expansion method (and calculation error) differed between 1986-1991 vs 1992 and 1994, resulting in the original overestimation of [n_returned] in 1986-1991 and underestimation in 1992 and 1994.